+++ pricing/RELEASES.md
@@ -55,3 +55,11 @@ Based on:
 - Speakeasy CLI 1.368.0 (2.399.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.2.3] pricing
+
+## 2024-08-20 00:25:53
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/pricing-api.yaml
+- Speakeasy CLI 1.376.0 (2.402.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.3.0] pricing